Finished with model changes.
SurveyResponse endpoints started

+++ PollPulse/PollPulse.Entities/Models/User.cs
@@ -3,9 +3,9 @@ using PollPulse.Entities.Models.Base;
 
 namespace PollPulse.Entities.Models
 {
-    public class User : IdentityUser<int>, IEntity
+    public class User : IdentityUser<long>, IEntity
     {
-        public Guid Guid { get; set; }
+        public Guid Guid { get; set; } 
         public string FirstName { get; set; }
         public string LastName { get; set; }
